server/requirements.txt
fastapi
uvicorn
pyserial
RPi.GPIO
python-socketio
uvicorn[standard]
